+++ contrib/cfran/model/cf-unit.h
@@ -2,12 +2,16 @@
 #define CF_UNIT_NET_DEVICE_H
 
 #include <ns3/cf-model.h>
+#include <ns3/cf-application.h>
 #include <ns3/net-device.h>
+
 
 namespace ns3
 {
 
 class Node;
+// struct CfModel;
+// struct UeTaskModel;
 
 /**
  * \defgroup cfran Models
@@ -46,49 +50,49 @@ class CfUnit : public Object
      * \param id the id of user
      * \param ueTask the model of this task
      */
-    void AddNewUeTaskForSchedule(uint16_t id, UeTaskModel ueTask);
+    void AddNewUeTaskForSchedule(uint64_t id, UeTaskModel ueTask);
 
     /**
      * \brief Allocate computing force for one task according to command and execute it
      * \param id the id of user
      * \param ueTask the model of this task
      */
-    void ScheduleNewTaskWithCommand(uint16_t id, UeTaskModel ueTask);
+    void ScheduleNewTaskWithCommand(uint64_t id, UeTaskModel ueTask);
 
     /**
      * \brief Allocate computing force by modifying  allcation map
      * \param id the id of user
      * \param ueTask the model of this task
      */
-    void AllocateCf(uint16_t id, UeTaskModel ueTask, CfModel cfModel);
+    void AllocateCf(uint64_t id, UeTaskModel ueTask, CfModel cfModel);
 
     /**
      * \brief Execute a specific task for the theoretical time
      * \param id the id of user
      * \param ueTask the model of this task
      */
-    void ExecuteTask(uint16_t id, UeTaskModel ueTask);
+    void ExecuteTask(uint64_t id, UeTaskModel ueTask);
 
     /**
      * \brief End a specific task after the theoretical time
      * \param id the id of user
      * \param ueTask the model of this task
      */
-    void EndTask(uint16_t id, UeTaskModel ueTask);
+    void EndTask(uint64_t id, UeTaskModel ueTask);
 
     /**
      * \brief Freeup the computing force by modifying allocation map
      * \param id the id of user
      * \param ueTask the model of this task
      */
-    void FreeCf(uint16_t id, UeTaskModel ueTask);
+    void FreeCf(uint64_t id, UeTaskModel ueTask);
 
     /**
      * \brief Delete the info of the task by modifying UE task map
      * \param id the id of user
      * \param ueTask the model of this task
      */
-    void DeleteUeTask(uint16_t id, UeTaskModel ueTask);
+    void DeleteUeTask(uint64_t id, UeTaskModel ueTask);
 
     /**
      * \brief Output the computing force allocation map
@@ -103,14 +107,14 @@ class CfUnit : public Object
 
   private:
     Ptr<Node> m_node; ///< the node
-    uint16_t m_id;    ///< the id of computing force unit
+    uint64_t m_id;    ///< the id of computing force unit
     CfModel m_cf;     ///< the total computing force of cfunit
     CfModel m_freeCf; ///< the free computing force
-    // std::map<uint16_t, std::vector<UeTaskModel>> m_ueTask; ///< the tasks of UEs, id -> tasks
+    // std::map<uint64_t, std::vector<UeTaskModel>> m_ueTask; ///< the tasks of UEs, id -> tasks
     ///< id of UE -> id of task -> ueTaskModel
-    std::map<uint16_t, std::map<uint16_t, UeTaskModel>> m_ueTask;
+    std::map<uint64_t, std::map<uint64_t, UeTaskModel>> m_ueTask;
     ///< id of UE -> id of task -> computing force allocated
-    std::map<uint16_t, std::map<uint16_t, CfModel>> m_cfAllocation;
+    std::map<uint64_t, std::map<uint64_t, CfModel>> m_cfAllocation;
 
     bool m_enableAutoSchedule; ///< schedule computing force autonomously of receieve command
 };
